Order players by their standing in the tournament.
Players are ordered by these criteria:
  * match points
  * opponents' match-win percentage
  * game-win percentage
  * opponents' game-win percentage

Definition of match-win percentage, from the DCI rules:
    "A player's match-win percentage is that player's accumulated match
    points divided by 3 times the number of rounds in which he or she
    competed, or 0.33, whichever is greater (Establishing a minimum
    match-win percentage [0.33] limits the effect low performances have
    when calculating and comparing opponents' match-win percentages in
    Step 1 of the tiebreakers)."
In other words, it's (matches won/matches played), plus a third of a
point for draws. There's a minimum of 0.33, which puts decreased weight
on this tiebreaker step when a player happens to have been seeded
against very weak opponents.

Definition of game-win percentage:
    "Similar to the match-win percentage, a player's game-win percentage
    is the total number of game points he or she earned divided by 3
    times the number of games played."
So, it's (games_won + (games_drawn/3))/games_played

As with match-win-percentage, there is a floor on opponents' game-win-
percentage, to limit the negative impact of a very weak seed.